[SQUASHED acff59a46c2 - f50e2bd3cab] Add General replicated distribution spec
When a query contains non deterministic operators on top of a replicated
distribution of tuples, the resulting distribution produced by these operators
is no longer replicated. Examples of such operators are global Limit, Window
Aggs, non trivial Aggs. They produce a "Tainted" form of replicated
distribution. ORCA does not have a way to represent this type of distribution.
In order to support this we introduce a "GeneralReplicated" distribution spec
and a "TaintedReplicated" distribution spec.
This commit introduces the "GeneralReplicated" distribution spec which is
reuqired only and cannot be derived. This distribution can only be satisfied by
a pure Replicated distribution or TaintedReplicated distribution.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>

Fixups in backport:
  - no GPOS_DEBUG macro
  - mdps have changed..
  - enum class new in C++11 (doesn't exist in 5X)